little test

+++ test.c
@@ -6,7 +6,7 @@
 /*   By: mjadid <mjadid@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/10 02:25:22 by mjadid            #+#    #+#             */
-/*   Updated: 2024/09/10 02:25:22 by mjadid           ###   ########.fr       */
+/*   Updated: 2024/09/13 05:08:00 by mjadid           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ void *thread_routine(void *data)
  // The pthread_self() function provides
  // this thread's own ID.
  tid = pthread_self();
- printf("%sThread [%ld]: The heaviest burden is to exist without living.%s\n",
-  YELLOW, tid, NC);
+ printf("%sThread [%lu]: The heaviest burden is to exist without living.%s\n",
+  YELLOW, (unsigned long)tid, NC);
  return (NULL); // The thread ends here.
 }
 
@@ -38,15 +38,19 @@ int main(void)
  // Creating the first thread that will go
  // execute its thread_routine function.
  pthread_create(&tid1, NULL, thread_routine, NULL);
- printf("Main: Created first thread [%ld]\n", tid1);
+ printf("Main: Created first thread [%lu]\n", (unsigned long)tid1);
  // Creating the second thread that will also execute thread_routine.
  pthread_create(&tid2, NULL, thread_routine, NULL);
- printf("Main: Created second thread [%ld]\n", tid2);
+ printf("Main: Created second thread [%lu]\n", (unsigned long)tid2);
  // The main thread waits for the new threads to end
  // with pthread_join.
  pthread_join(tid1, NULL);
- printf("Main: Joining first thread [%ld]\n", tid1);
+ printf("Main: Joining first thread [%lu]\n", (unsigned long)tid1);
  pthread_join(tid2, NULL);
- printf("Main: Joining second thread [%ld]\n", tid2);
+ printf("Main: Joining second thread [%lu]\n", (unsigned long)tid2);
  return (0);
 }
+
+
+
+
